error/app/requirements.txt
azure-ai-projects~=1.0.0b11
azure-identity~=1.20.0
ansible-core~=2.17.0
azure-functions
azure-ai-projects>=1.0.0b11
azure-search-documents
azure-storage-queue
azure-core
pydantic